WeConnect-main/WeConnect-main/backend/requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.47.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-heroku==0.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.3
filelock==3.4.0
fonttools==4.28.5
gunicorn==20.1.0
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.4
nose==1.3.7
notebook==6.4.6
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
sympy==1.9
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.0.0
widgetsnbextension==3.5.2
yahoo-finance==1.4.0
